[Silver IV] Title: 파스칼 삼각형, Time: 0 ms, Memory: 2024 KB -BaekjoonHub

+++ 백준/Silver/15489. 파스칼 삼각형/파스칼 삼각형.cc
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 #define MAX 31
-long long d[MAX][MAX];
+int d[MAX][MAX];
 int main()
 {
     cin.tie(NULL);
